Zomato Dataset Analysis and Visualization with Python
The "Zomato Dataset Analysis and Visualization" project is a Python-based data analysis project that aims to explore and visualize the Zomato restaurant dataset. The Zomato dataset is a collection of restaurant ratings and reviews from around the world, and this project uses various data analysis and visualization techniques to gain insights into the data.

The project involves importing and cleaning the dataset, performing exploratory data analysis, and creating various visualizations using libraries such as matplotlib, seaborn, and plotly. The visualizations include bar charts, scatter plots, heatmaps, and interactive visualizations using plotly.

The insights gained from this analysis and visualization project can be useful for restaurant owners, food bloggers, and food enthusiasts to understand the trends and patterns in the restaurant industry. Additionally, this project demonstrates the use of Python for data analysis and visualization and can serve as a learning resource for beginners to data science and Python programming.